Add files via upload
I have used HTML, CSS, and JavaScript to create this project. html CSS, I have used little transition animation. This project is mainly used for displaying things in our daily lives, like price of food and quantity of items and it calculate and shows the total cost of the food